capability: unify the capability registry across all use-case families
Extract the duplicated add_capability / get_capability / run_capability_json /
run_capability_simple_text / get_capability_block logic into a single
CapabilityRegistry mixin (capability.py). Agent and CapabilityManager now both
inherit it, so capability management has one implementation shared by every
use-case family: the web agents and MinimalToolCallPrivEscLinux (Agent), and
the CommandStrategy priv-esc use-cases and the SimpleStrategy web-api engines
(CapabilityManager). No use-case files change - they keep calling the same
methods.

Also fixes CapabilityManager's class-level mutable-default (_capabilities/
_default_capability shared across instances) by storing them per instance, and
gives Agent the run_capability_simple_text that TemplatedAgent already calls.

Behaviour-preserving: the ChatAgent characterization and the full suite stay
green.
